Add ClientInterface and ServerInterface classes, which replaces former interfaces. Add deprecation notice to Client and Server interfaces. They will be removed in 3.0 to match laminas-coding-standard.
