+++ src/Client.php
@@ -10,17 +10,8 @@ namespace Laminas\Server;
 
 /**
  * Client Interface
+ * @deprecated Client is replaced by ClientInterface and will be removed in 3.0.
  */
-interface Client
+interface Client extends ClientInterface
 {
-    /**
-     * Executes remote call
-     *
-     * Unified interface for calling custom remote methods.
-     *
-     * @param  string $method Remote call name.
-     * @param  array $params Call parameters.
-     * @return mixed Remote call results.
-     */
-    public function call($method, $params = []);
 }
